Librarized smoke_test and made ai_observer Python.
